# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger RPC POC Description",
        "title": "Swagger RPC POC",
        "contact": {},
        "version": "0.1.0"
    },
    "paths": {
        "/Method1": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POC"
                ],
                "summary": "Method1 handler.",
                "parameters": [
                    {
                        "description": "Request Method1",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Method1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Method1",
                        "schema": {
                            "$ref": "#/definitions/GoodResponse"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/Method2": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POC"
                ],
                "summary": "Method2 handler.",
                "parameters": [
                    {
                        "description": "Request Method2",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Method2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Method2",
                        "schema": {
                            "$ref": "#/definitions/GoodResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "GoodResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "Method1": {
            "type": "object",
            "properties": {
                "field1": {
                    "type": "string",
                    "example": "field1"
                },
                "field2": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "Method2": {
            "type": "object",
            "properties": {
                "field1": {
                    "type": "boolean",
                    "example": true
                },
                "field2": {
                    "type": "string",
                    "example": "field2"
                },
                "field3": {
                    "type": "integer",
                    "example": 2
                }
            }
        }
    }
}